FacialAPI/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
numpy==1.26.3
face-recognition==1.3.0
Pillow==10.2.0
pydantic==2.5.3
